DSPy/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via litellm
aiosignal==1.4.0
    # via aiohttp
alembic==1.16.4
    # via optuna
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   asyncer
    #   dspy
    #   httpx
    #   openai
asyncer==0.0.8
    # via dspy
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via dspy
cachetools==6.1.0
    # via dspy
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via litellm
cloudpickle==3.1.1
    # via dspy
colorlog==6.9.0
    # via optuna
diskcache==5.6.3
    # via dspy
distro==1.9.0
    # via openai
dspy==2.6.27
    # via dspy-examples (pyproject.toml)
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   litellm
    #   openai
huggingface-hub==0.33.4
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via dspy
json-repair==0.47.8
    # via dspy
jsonschema==4.24.1
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.74.4
    # via dspy
magicattr==0.1.6
    # via dspy
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
numpy==2.3.1
    # via
    #   dspy
    #   optuna
openai==1.97.0
    # via
    #   dspy
    #   litellm
optuna==4.4.0
    # via dspy
packaging==25.0
    # via
    #   huggingface-hub
    #   optuna
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   dspy
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
python-dotenv==1.1.1
    # via litellm
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   optuna
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   dspy
    #   tiktoken
requests==2.32.4
    # via
    #   dspy
    #   huggingface-hub
    #   tiktoken
rich==14.0.0
    # via dspy
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   alembic
    #   optuna
tenacity==9.1.2
    # via dspy
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.2
    # via litellm
tqdm==4.67.1
    # via
    #   dspy
    #   huggingface-hub
    #   openai
    #   optuna
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
ujson==5.10.0
    # via dspy
urllib3==2.5.0
    # via requests
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
